% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6core.R
\name{Lithosphere}
\alias{Lithosphere}
\title{R6 Class representing the fundamental framework for building a geological 3D-model in Lithosphere}
\description{
A Lithosphere-Object has, in xy-direction, a minimum and a maximum coordinate and a resolution as well as a predefined stratigraphic sequence
}
\details{
Once the model is initialized with the basic grid constraints, provide well markers indicating lithology changes and (optionally) a dem and start the interpolation process.
}
\examples{

data("synthetic_dem_1") 
data("synthetic_welldata_1")

mymodel <- Lithosphere$new(
xmin = 0, xmax = 1000, xres = 10, 
ymin = 0, ymax = 1000, yres = 10, 
stratigraphy = c("upper_sand", "intermediate_clay", "lower_marl")
)

mymodel$set_DEM(synthetic_dem_1)
mymodel$set_markers(xyzg = synthetic_welldata_1[, c("x", "y", "z", "stratigraphy")])
mymodel$interpolate_surfaces()

mymodel$export_to_data_frame() 


}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{idp}}{inverse distance power}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Lithosphere$new()}}
\item \href{#method-print}{\code{Lithosphere$print()}}
\item \href{#method-set_DEM}{\code{Lithosphere$set_DEM()}}
\item \href{#method-set_markers}{\code{Lithosphere$set_markers()}}
\item \href{#method-interpolate_surfaces}{\code{Lithosphere$interpolate_surfaces()}}
\item \href{#method-export_to_data_frame}{\code{Lithosphere$export_to_data_frame()}}
\item \href{#method-plot3d}{\code{Lithosphere$plot3d()}}
\item \href{#method-clone}{\code{Lithosphere$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new lithosphere model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lithosphere$new(
  xmin,
  xmax,
  xres,
  ymin,
  ymax,
  yres,
  stratigraphy,
  verbose = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{xmin}}{minimum x coordinate (left border). Will be truncated to integer.}

\item{\code{xmax}}{maximum x coordinate (right border). Will be truncated to integer.}

\item{\code{xres}}{numeric vector of length 1 to set the horizontal resolution. Will be truncated to integer.}

\item{\code{ymin}}{minimum y coordinate (lower border). Will be truncated to integer.}

\item{\code{ymax}}{maximum y coordinate (upper border). Will be truncated to integer.}

\item{\code{yres}}{numeric vector of length 1 to set the vertical resolution. Will be truncated to integer.}

\item{\code{stratigraphy}}{character vector with names of the stratigraphic sequence (the layers) that shall be modeled sorted from the uppermost to the lowermost unit. Getting the correct order here ist really important!}

\item{\code{verbose}}{logical. Print out what you ordered?}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
The models printing method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lithosphere$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_DEM"></a>}}
\if{latex}{\out{\hypertarget{method-set_DEM}{}}}
\subsection{Method \code{set_DEM()}}{
Set the models uppermost border. Usually thought to be a digital elevation model (DEM) representing the ground level.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lithosphere$set_DEM(xyz)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{xyz}}{matrix or data.frame with exactly three columns: x, y and z coordinates.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_markers"></a>}}
\if{latex}{\out{\hypertarget{method-set_markers}{}}}
\subsection{Method \code{set_markers()}}{
Add well markers
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lithosphere$set_markers(xyzg)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{xyzg}}{a \code{data.frame} with four columns: the x, the y and the z coordinate of a marker point as well as a label corresponding to the stratigraphy as defined at model initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-interpolate_surfaces"></a>}}
\if{latex}{\out{\hypertarget{method-interpolate_surfaces}{}}}
\subsection{Method \code{interpolate_surfaces()}}{
Start the interpolation and possibly the subsequent erosion of all stratigraphic layers within a model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lithosphere$interpolate_surfaces(engine = "uk", erosion = "topdown")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{engine}}{a character vector of length 1 stating the interpolation algorithm used to transform markers into continuous surfaces. Currently available are "idw" (inverse distance weighted), "ok" (ordinary kriging), and "uk" (universal kriging - the default).}

\item{\code{erosion}}{a character vector of length 1 stating the type of erosion applied to each layer (with the exception of the DEM) after interpolation. Available options are "topdown" (the default), "bottomup" and "none" (which will not cut any intersecting surfaces at all).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-export_to_data_frame"></a>}}
\if{latex}{\out{\hypertarget{method-export_to_data_frame}{}}}
\subsection{Method \code{export_to_data_frame()}}{
Return a data.frame of interpolated surface coordinates
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lithosphere$export_to_data_frame(shape = "long")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{shape}}{a character vector of length 1: either "long" (the default) or "wide".}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{data.frame} formated either in line with \code{\link[tidyr]{pivot_wider}} or \code{\link[tidyr]{pivot_longer}}, depending on which format was specified via the \code{shape} argument.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot3d"></a>}}
\if{latex}{\out{\hypertarget{method-plot3d}{}}}
\subsection{Method \code{plot3d()}}{
Plot a 3D-representation of the model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lithosphere$plot3d(surface_colors = NA, exag_fct = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{surface_colors}}{A character vector of color names as accepted by R. Ideally, the length of the vector should correspond to the number of stratigraphic units (+DEM) of the model but selecting any other number of colors will be dealt with, too.}

\item{\code{exag_fct}}{numeric specifying the exaggeration factor of the visualization in z-direction. Defaults to 1 but maybe try 3.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lithosphere$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
