% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{synthetic_dem_1}
\alias{synthetic_dem_1}
\title{A simulated digital elevation model}
\format{
\describe{
\item{x}{x coordinate}
\item{y}{y coordinate}
\item{z}{z coordinate}
}
}
\source{
made with \code{\link[NLMR]{nlm_fbm}}, see \href{https://ropensci.github.io/NLMR/}{NLMR website}
}
\usage{
synthetic_dem_1
}
\description{
A dataset mimicking a DEM of ground surface data to serve as an example for calculations in \code{lithosphere}. Resolution is 10.
}
\keyword{datasets}
