% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{synthetic_welldata_1}
\alias{synthetic_welldata_1}
\title{A set of simulated well markers}
\format{
An object of class \code{data.frame} with 150 rows and 5 columns.
}
\source{
made with \code{\link[NLMR]{NLMR-package}}, see \href{https://ropensci.github.io/NLMR/}{NLMR website}
}
\usage{
synthetic_welldata_1
}
\description{
A dataset of 50 fictional wells and associated stratigraphy to serve as an example for calculations in \code{lithosphere}
#' @format
\describe{
\item{well_name}{a fictional name for the well to serve as unique identifier}
\item{x}{x coordinate}
\item{y}{y coordinate}
\item{z}{z coordinate of the bottom of a stratigraphic unit}
\item{stratigraphy}{a unified geological description of the lithology at xyz}
}
}
\keyword{datasets}
